+++ src/groovy-lint.js
@@ -4,6 +4,7 @@ const DEFAULT_VERSION = "3.0.0";
 
 // Imports
 const debug = require("debug")("npm-groovy-lint");
+const performance = require("perf_hooks").performance;
 
 const NpmGroovyLintFix = require("./groovy-lint-fix");
 const CodeNarcCaller = require("./codenarc-caller");
@@ -19,6 +20,7 @@ class NpmGroovyLint {
     args = []; // Command line arguments
 
     // Internal
+    origin = "initialCall";
     jdeployFile;
     jdeployFilePlanB;
     jdeployRootPath;
@@ -52,6 +54,7 @@ class NpmGroovyLint {
         this.jdeployFilePlanB = internalOpts.jdeployFilePlanB || process.env.JDEPLOY_FILE_PLAN_B || "originaljdeployPlanB.js";
         this.jdeployRootPath = internalOpts.jdeployRootPath || process.env.JDEPLOY_ROOT_PATH || __dirname;
         this.parseOptions = internalOpts.parseOptions !== false;
+        this.origin = internalOpts.origin || this.origin;
     }
 
     // Run linting (and fixing if --fix)
@@ -72,12 +75,16 @@ class NpmGroovyLint {
         debug(`Fix errors for ${JSON.stringify(errorIds)} on existing NpmGroovyLint instance`);
         const codeNarcFactoryResult = await prepareCodeNarcCall(this.options, this.jdeployRootPath);
         this.setMethodResult(codeNarcFactoryResult);
-        this.fixer = new NpmGroovyLintFix(this.lintResult, {
-            verbose: optns.verbose || this.options.verbose,
-            fixrules: optns.fixrules || this.options.fixrules,
-            source: optns.source || this.options.source,
-            save: this.tmpGroovyFileName ? false : true
-        });
+        this.fixer = new NpmGroovyLintFix(
+            this.lintResult,
+            {
+                verbose: optns.verbose || this.options.verbose,
+                fixrules: optns.fixrules || this.options.fixrules,
+                source: optns.source || this.options.source,
+                save: this.tmpGroovyFileName ? false : true
+            },
+            { origin: "externalCallToFix" }
+        );
         await this.fixer.run({ errorIds: errorIds, propagate: true });
         this.lintResult = this.fixer.updatedLintResult;
         // Lint again after fix if requested (for the moment we prefer to trigger that from VsCode, for better UX)
@@ -158,6 +165,7 @@ class NpmGroovyLint {
 
         // Kill running CodeNarcServer
         if (this.options.killserver) {
+            const startPerf = performance.now();
             const codeNarcCaller = new CodeNarcCaller(this.codenarcArgs, this.serverStatus, this.args, this.options, {
                 jdeployFile: this.jdeployFile,
                 jdeployFilePlanB: this.jdeployFilePlanB,
@@ -165,6 +173,7 @@ class NpmGroovyLint {
             });
             this.outputString = await codeNarcCaller.killCodeNarcServer();
             console.info(this.outputString);
+            this.manageStats(startPerf);
             return false;
         }
 
@@ -182,6 +191,7 @@ class NpmGroovyLint {
         - Call CodeNarc using org.codenarc.CodeNarc (originaljdeployPlanB.js)
     */
     async callCodeNarc() {
+        const startPerf = performance.now();
         let serverCallResult = { status: null };
         const codeNarcCaller = new CodeNarcCaller(this.codenarcArgs, this.serverStatus, this.args, this.options, {
             jdeployFile: this.jdeployFile,
@@ -195,6 +205,7 @@ class NpmGroovyLint {
             serverCallResult = await codeNarcCaller.callCodeNarcJava();
         }
         this.setMethodResult(serverCallResult);
+        this.manageStats(startPerf);
     }
 
     // After CodeNarc call
@@ -222,8 +233,13 @@ class NpmGroovyLint {
                 });
                 await this.fixer.run();
                 this.lintResult = this.fixer.updatedLintResult;
+                // Post actions
+                const checkIfFixAgainRequiredRes = this.checkIfFixAgainRequired();
+                if (checkIfFixAgainRequiredRes.runAgain === true && this.origin !== "fixAgainAfterFix") {
+                    await this.fixAgainAfterFix(checkIfFixAgainRequiredRes.files);
+                }
                 // If there has been fixes, call CodeNarc again to get updated error list
-                if (this.fixer.fixedErrorsNumber > 0 && this.options.nolintafter !== true) {
+                if (this.fixer.fixedErrorsNumber > 0 && this.options.nolintafter !== true && this.origin !== "fixAgainAfterFix") {
                     await this.lintAgainAfterFix();
                 }
             }
@@ -235,25 +251,74 @@ class NpmGroovyLint {
         manageDeleteTmpFiles(this.tmpGroovyFileName, this.tmpRuleSetFileName);
     }
 
+    // Check if fixed errors required a new lint & fix
+    checkIfFixAgainRequired() {
+        let runAgain = false;
+        const runAgainOnFiles = {};
+        for (const file of Object.keys(this.lintResult.files)) {
+            for (const err of this.lintResult.files[file].errors) {
+                if (err.fixed === true && err.triggersAgainAfterFix && err.triggersAgainAfterFix.length > 0) {
+                    runAgain = true;
+                    runAgainOnFiles[file] = runAgainOnFiles[file] ? runAgainOnFiles[file] : { fixrules: [] };
+                    runAgainOnFiles[file].fixrules.push(...err.triggersAgainAfterFix);
+                }
+            }
+        }
+        return { runAgain: runAgain, files: runAgainOnFiles };
+    }
+
     // Lint again after fixes and merge in existing results
     async lintAgainAfterFix() {
         // same Options except fix = false & output = none
         const lintAgainOptions = JSON.parse(JSON.stringify(this.options));
-        debug(`Fix is done, lint again with options ${JSON.stringify(lintAgainOptions)}`);
         if (this.options.source) {
             lintAgainOptions.source = this.lintResult.files[0].updatedSource;
         }
         lintAgainOptions.fix = false;
         lintAgainOptions.output = "none";
+        debug(`Fix is done, lint again with options ${JSON.stringify(lintAgainOptions)}`);
         const newLinter = new NpmGroovyLint(lintAgainOptions, {
             parseOptions: false,
             jdeployFile: this.jdeployFile,
-            jdeployRootPath: this.jdeployRootPath
+            jdeployRootPath: this.jdeployRootPath,
+            origin: "lintAgainAfterFix"
         });
         // Run linter
         await newLinter.run();
         // Merge new linter results in existing results
         this.lintResult = this.mergeResults(this.lintResult, newLinter.lintResult);
+    }
+
+    // Fix again after fix because fixed rules contained triggersAgainAfterFix property (for the moment, only Indentation rule)
+    async fixAgainAfterFix(filesAndRulesToProcess) {
+        const fixAgainOptions = JSON.parse(JSON.stringify(this.options));
+        // Gather rules to lint & fix again
+        let fixRules = [];
+        for (const file of Object.keys(filesAndRulesToProcess)) {
+            fixRules.push(...filesAndRulesToProcess[file].fixrules);
+        }
+        fixRules = [...new Set(fixRules)]; // Remove duplicates
+        if (this.options.source) {
+            fixAgainOptions.source = this.lintResult.files[0].updatedSource;
+        }
+        // Lint & fix again only the requested rules for better perfs
+        delete fixAgainOptions.format;
+        delete fixAgainOptions.rules;
+        fixAgainOptions.rulesets = fixRules.join(",");
+        fixAgainOptions.fix = true;
+        fixAgainOptions.fixrules = fixRules.join(",");
+        fixAgainOptions.output = "none";
+        // Process lint & fix
+        debug(`Fix triggered rule necessing another lint & fix, do it again with options ${JSON.stringify(fixAgainOptions)}`);
+        const newLinter = new NpmGroovyLint(fixAgainOptions, {
+            jdeployFile: this.jdeployFile,
+            jdeployRootPath: this.jdeployRootPath,
+            origin: "fixAgainAfterFix"
+        });
+        await newLinter.run();
+
+        // Merge new linter & fixer results in existing results
+        this.lintResult = this.mergeFixAgainResults(this.lintResult, newLinter.lintResult);
     }
 
     // Merge results after control lint after fixing
@@ -275,10 +340,38 @@ class NpmGroovyLint {
         return mergedLintResults;
     }
 
+    // Merge results after second fix performed (only get updated source)
+    mergeFixAgainResults(lintResToUpdate, lintResAfterNewFix) {
+        for (const afterNewFixResFileNm of Object.keys(lintResAfterNewFix.files)) {
+            // Set updatedSource in results in provided
+            if (lintResAfterNewFix.files[afterNewFixResFileNm].updatedSource) {
+                lintResToUpdate.files[afterNewFixResFileNm].updatedSource = lintResAfterNewFix.files[afterNewFixResFileNm].updatedSource;
+            }
+        }
+        return lintResToUpdate;
+    }
+
     // Set lib results on this NpmGroovyLint instance
     setMethodResult(libResult) {
         for (const propName of Object.keys(libResult)) {
             this[propName] = libResult[propName];
+        }
+    }
+
+    // Increment stats for test classes in necessary
+    manageStats(startPerf) {
+        if (globalThis && globalThis.codeNarcCallsCounter >= 0) {
+            globalThis.codeNarcCallsCounter++;
+            const optionsLog = JSON.parse(JSON.stringify(this.options));
+            for (const prop of ["source", "rules", "verbose", "loglevel", "serverhost", "serverport", "_"]) {
+                delete optionsLog[prop];
+            }
+            globalThis.codeNarcCalls.push({
+                origin: this.origin,
+                elapse: parseInt(performance.now() - startPerf),
+                options: optionsLog,
+                args: this.codenarcArgs
+            });
         }
     }
 }