+++ test/lint-api.test.js
@@ -3,15 +3,17 @@
 const NpmGroovyLint = require('../src/groovy-lint.js');
 let assert = require('assert');
 const fse = require("fs-extra");
+const { beforeEachTestCase, checkCodeNarcCallsCounter, SAMPLE_FILE_SMALL, SAMPLE_FILE_SMALL_PATH } = require('./helpers/common');
 
 describe('Lint with API', () => {
+    beforeEach(beforeEachTestCase);
 
     it('(API:file) should generate text console output', async () => {
         const linter = await new NpmGroovyLint([
             process.execPath,
             '',
             '--path', '"jdeploy-bundle/lib/example"',
-            '--files', '**/SampleFile.groovy',
+            '--files', '**/' + SAMPLE_FILE_SMALL,
             '--verbose'
         ], {
             jdeployRootPath: 'jdeploy-bundle',
@@ -21,6 +23,7 @@ describe('Lint with API', () => {
         assert(linter.outputString.includes('warning'), 'Output string contains warning');
         assert(linter.lintResult.summary.totalFoundWarningNumber > 0, 'Warnings found');
         assert(linter.lintResult.summary.totalFoundInfoNumber > 0, 'Infos found');
+        checkCodeNarcCallsCounter(1);
     });
 
     it('(API:file) should generate json output with rules', async () => {
@@ -28,7 +31,7 @@ describe('Lint with API', () => {
             process.execPath,
             '',
             '--path', '"jdeploy-bundle/lib/example"',
-            '--files', '**/SampleFile.groovy',
+            '--files', '**/' + SAMPLE_FILE_SMALL,
             '--output', 'json',
             '--loglevel', 'warning'
         ],
@@ -36,6 +39,7 @@ describe('Lint with API', () => {
         assert(linter.status === 0, 'Status is 0');
         assert(linter.outputString.includes(`"totalFoundWarningNumber":`), 'Property totalFoundWarningNumber is in result');
         assert(linter.lintResult.summary.totalFoundWarningNumber > 0, 'Warnings found');
+        checkCodeNarcCallsCounter(1);
     });
 
     it('(API:file) should generate codenarc HTML file report', async () => {
@@ -44,13 +48,14 @@ describe('Lint with API', () => {
             process.execPath,
             '',
             '--path', 'jdeploy-bundle/lib/example',
-            '--files', '**/Jenkinsfile',
+            '--files', '**/' + SAMPLE_FILE_SMALL,
             '--output', reportFileName],
             { jdeployRootPath: 'jdeploy-bundle' }).run();
 
         assert(linter.status === 0, 'Status is 0');
         assert(fse.existsSync(reportFileName), 'CodeNarc HTML report generated');
         fse.removeSync(reportFileName);
+        checkCodeNarcCallsCounter(1);
     });
 
     it('(API:file) should generate codenarc XML file report', async () => {
@@ -59,13 +64,14 @@ describe('Lint with API', () => {
             process.execPath,
             '',
             '--path', 'jdeploy-bundle/lib/example',
-            '--files', '**/Jenkinsfile',
+            '--files', '**/' + SAMPLE_FILE_SMALL,
             '--output', reportFileName],
             { jdeployRootPath: 'jdeploy-bundle' }).run();
 
         assert(linter.status === 0, 'Status is 0');
         assert(fse.existsSync(reportFileName), 'CodeNarc XML report generated');
         fse.removeSync(reportFileName);
+        checkCodeNarcCallsCounter(1);
     });
 
     it('(API:file) should use --codenarcargs to generate XML report', async () => {
@@ -83,6 +89,7 @@ describe('Lint with API', () => {
         assert(linter.status === 0, 'Status is 0');
         assert(fse.existsSync(reportFileName), 'XML CodeNarc report has been generated');
         fse.removeSync(reportFileName);
+        checkCodeNarcCallsCounter(1);
     });
 
     it('(API:file) should run on a Jenkinsfile', async () => {
@@ -90,7 +97,7 @@ describe('Lint with API', () => {
             process.execPath,
             '',
             '--path', '"jdeploy-bundle/lib/example"',
-            '-f', '"**/Jenkinsfile"',
+            '-f', '**/Jenkinsfile',
             '--verbose'
         ], {
             jdeployRootPath: 'jdeploy-bundle',
@@ -100,6 +107,7 @@ describe('Lint with API', () => {
         assert(linter.outputString.includes('warning'), 'Output string contains warning');
         assert(linter.lintResult.summary.totalFoundWarningNumber > 0, 'Warnings found');
         assert(linter.lintResult.summary.totalFoundInfoNumber > 0, 'Infos found');
+        checkCodeNarcCallsCounter(1);
     });
 
     it('(API:Server) should kill running server', async () => {
@@ -115,6 +123,7 @@ describe('Lint with API', () => {
         if (process.platform.includes('win')) { // Doesn't work in Linux yet :/ 
             assert(linter.status === 0, 'Status is 0');
             assert(linter.outputString.includes('CodeNarcServer terminated'), 'CodeNarcServer has been terminated');
+            checkCodeNarcCallsCounter(1);
         }
         else {
             console.log('Test (API:Server) should kill running server skipped: CodeNarcServer works only on windows for now , and we are on ' + process.platform);
@@ -133,6 +142,7 @@ describe('Lint with API', () => {
         }).run();
         assert(linter.status === 0, 'Status is 0');
         assert(linter.outputString.includes('CodeNarcServer was not running'), 'CodeNarcServer not killed because not running');
+        checkCodeNarcCallsCounter(1);
     });
 
     it('(API:help) should show npm-groovy-lint help', async () => {
@@ -180,14 +190,14 @@ describe('Lint with API', () => {
         }).run();
         assert(linter.status === 0, 'Status is 0');
         assert(linter.codeNarcStdOut.includes('where OPTIONS are zero or more command-line options'), 'CodeNarc help is displayed');
+        checkCodeNarcCallsCounter(1);
     });
 
 
     it('(API:source) should run with source only', async () => {
-        const sampleFilePath = 'lib/example/SampleFile.groovy';
         const npmGroovyLintConfig = {
-            source: fse.readFileSync(sampleFilePath).toString(),
-            sourcefilepath: sampleFilePath,
+            source: fse.readFileSync(SAMPLE_FILE_SMALL_PATH).toString(),
+            sourcefilepath: SAMPLE_FILE_SMALL_PATH,
             output: 'txt',
             verbose: true
         };
@@ -197,13 +207,13 @@ describe('Lint with API', () => {
         }).run();
         assert(linter.status === 0, 'Status is 0');
         assert(linter.lintResult.files[0].errors.length > 0, 'Errors have been found');
+        checkCodeNarcCallsCounter(1);
     });
 
     it('(API:source) should run without CodeNarc Server', async () => {
-        const sampleFilePath = 'lib/example/SampleFile.groovy';
         const npmGroovyLintConfig = {
-            source: fse.readFileSync(sampleFilePath).toString(),
-            sourcefilepath: sampleFilePath,
+            source: fse.readFileSync(SAMPLE_FILE_SMALL_PATH).toString(),
+            sourcefilepath: SAMPLE_FILE_SMALL_PATH,
             noserver: true,
             output: 'none',
             verbose: true
@@ -214,6 +224,7 @@ describe('Lint with API', () => {
         }).run();
         assert(linter.status === 0, 'Status is 0');
         assert(linter.lintResult.files[0].errors.length > 0, 'Errors have been found');
+        checkCodeNarcCallsCounter(1);
     });
 
 });